Move issues.tsx out of test dir lol
